trial invoice email

+++ internal/logic/gateway/api/gateway_user.go
@@ -35,7 +35,7 @@ func QueryAndCreateChannelUser(ctx context.Context, gateway *entity.MerchantGate
 func QueryAndCreateChannelUserWithOutPaymentMethod(ctx context.Context, gateway *entity.MerchantGateway, userId uint64) *entity.GatewayUser {
 	gatewayUser := query.GetGatewayUser(ctx, userId, gateway.Id)
 	if gatewayUser == nil {
-		user := query.GetUserAccountById(ctx, uint64(userId))
+		user := query.GetUserAccountById(ctx, userId)
 		utility.Assert(user != nil, "user not found")
 		utility.Assert(len(user.Email) > 0, "invalid user email")
 		create, err := GetGatewayServiceProvider(ctx, gateway.Id).GatewayUserCreate(ctx, gateway, user)
